move checkpoint_sims to Checkpointer; delete user_logging.jl
Move checkpoint_sims function to the Checkpointer module and delete
the user_logging.jl file. This should not change any behavior.
